requirements.txt
Click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==0.15.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.4
python-dotenv==0.15.0